# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AnimeThemes.moe Api Documentation",
        "description": "AnimeThemes.moe Laravel RESTful API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/anime/": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Get paginated listing of Anime",
                "description": "Returns listing of Anime",
                "operationId": "getAnimes",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "The search query. Mapping is to anime.name and anime.synonyms.text.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "bakemonogatari"
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Comma-separated list of included related resources. Allowed list is synonyms, series, themes, themes.entries, themes.entries.videos, themes.song, themes.song.artists & externalResources.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "synonyms,series"
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter anime by year",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2009"
                    },
                    {
                        "name": "season",
                        "in": "query",
                        "description": "Filter anime by season. Case-insensitive options are Winter, Spring, Summer & Fall.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Summer"
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order anime by field. Case-insensitive options are anime_id, created_at, updated_at, alias, name, year & season.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "updated_at"
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "Direction of anime ordering. Case-insensitive options are asc & desc.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "desc"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The number of resources to return per page. Acceptable range is [1-100]. Default value is 100.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 50
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "The comma-separated list of fields by resource type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fields[anime]=name,alias"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "anime": {
                                            "description": "Display a listing of the resource.",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AnimeResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/anime/{alias}": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Get properties of Anime",
                "description": "Returns properties of Anime",
                "operationId": "getAnime",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Comma-separated list of included related resources. Allowed list is synonyms, series, themes, themes.entries, themes.entries.videos, themes.song, themes.song.artists & externalResources.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "synonyms,series"
                    },
                    {
                        "name": "fields",
                        "in": "query",
                        "description": "The comma-separated list of fields by resource type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "fields[anime]=name,alias"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnimeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found!"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}